+++ activesupport/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   Deprecate `config.active_support.remove_deprecated_time_with_zone_name`.
+
+    *Rafael Mendonça França*
+
+*   Remove deprecated override of `ActiveSupport::TimeWithZone.name`.
+
+    *Rafael Mendonça França*
+
 *   Deprecate `config.active_support.disable_to_s_conversion`.
 
     *Rafael Mendonça França*